+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 
 ## [v1.2.0](https://github.com/pestphp/defstudio-plugin-laravel-expectations/compare/v1.1.0...v1.1.0) - 2021-10-11
+### Added
+- `toExistInStorage()` expectation
+
 ### Fixed
 - [Pest IntelliJ](https://github.com/pestphp/pest-intellij) support for autocompletion